LIFT3D/requirements.txt
wandb[media]==0.19.0
imageio==2.23.0
imageio-ffmpeg==0.5.1
hydra-core==1.3.2
easydict==1.13
termcolor==2.4.0
opencv-python-headless==4.10.0.84
open3d==0.18.0
zarr==2.18.3
gymnasium==1.0.0
numpy==1.26.4
timm==0.9.10
ftfy==6.3.0
regex==2024.9.11
gdown==5.2.0
omegaconf==2.3.0
setuptools==59.5.0 
multimethod==1.12
shortuuid==1.0.13
numcodecs[msgpack]==0.13.1
einops==0.8.0
lightning==2.4.0
rich==13.7.1
huggingface-hub==0.26.5
